Revert "Story screen"

+++ source/core/src/main/com/deco2800/game/components/bridge/Location.java
@@ -1,7 +1,7 @@
 package com.deco2800.game.components.bridge;
 
 public interface Location {
-    float getTop();
-    float getMid();
-    float getBot();
+    int getTop();
+    int getMid();
+    int getBot();
 }
